HHH-12185 - Simplify SessionFactoryBuilder / SessionFactoryOptions handling
Option #1 - still building a stand-alone, immutable options object
