+++ hibernate-core/src/main/java/org/hibernate/hql/internal/ast/util/SessionFactoryHelper.java
@@ -419,6 +419,6 @@ public class SessionFactoryHelper {
 	}
 
 	public boolean isStrictJPAQLComplianceEnabled() {
-		return sfi.getSessionFactoryOptions().isStrictJpaQueryLanguageCompliance();
+		return sfi.getSessionFactoryOptions().getJpaCompliance().isJpaQueryComplianceEnabled();
 	}
 }
